requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai
faiss-cpu
pydantic
python-dotenv
beautifulsoup4
requests
streamlit
wikipedia
ipykernel
pypdf
